Update printer.cfg
Added load and unload filament macros to simulate factory controls, only better. You don't need to keep an eye on it. Press the button and it will wait to heat (to a much lower temperature), turn on LED when the extruder is going to move, load 550 or unload 500mm, turn off heaters and turn off LED.